+++ lib/sitemap-index-stream.ts
@@ -1,16 +1,5 @@
-import { promisify } from 'util';
-import { URL } from 'url';
-import { stat, createWriteStream } from 'fs';
-import { createGzip } from 'zlib';
-import {
-  Transform,
-  TransformOptions,
-  TransformCallback,
-  Writable,
-} from 'stream';
+import { Transform, TransformOptions, TransformCallback } from 'stream';
 import { IndexItem, SitemapItemLoose, ErrorLevel } from './types';
-import { UndefinedTargetFolder } from './errors';
-import { chunk } from './utils';
 import { SitemapStream, stylesheetInclude } from './sitemap-stream';
 import { element, otag, ctag } from './sitemap-xml';
 import { WriteStream } from 'fs';
@@ -21,7 +10,6 @@ export enum IndexTagNames {
   lastmod = 'lastmod',
 }
 
-const statPromise = promisify(stat);
 const xmlDec = '<?xml version="1.0" encoding="UTF-8"?>';
 
 const sitemapIndexTagStart =
@@ -79,87 +67,9 @@ export class SitemapIndexStream extends Transform {
   }
 }
 
-/**
- * Shortcut for `new SitemapIndex (...)`.
- * Create several sitemaps and an index automatically from a list of urls
- *
- * @deprecated Use SitemapAndIndexStream
- * @param   {Object}        conf
- * @param   {String|Array}  conf.urls
- * @param   {String}        conf.targetFolder where do you want the generated index and maps put
- * @param   {String}        conf.hostname required for index file, will also be used as base url for sitemap items
- * @param   {String}        conf.sitemapName what do you want to name the files it generats
- * @param   {Number}        conf.sitemapSize maximum number of entries a sitemap should have before being split
- * @param   {Boolean}       conf.gzip whether to gzip the files (defaults to true)
- * @return  {SitemapIndex}
- */
-export async function createSitemapsAndIndex({
-  urls,
-  targetFolder,
-  hostname,
-  sitemapName = 'sitemap',
-  sitemapSize = 50000,
-  gzip = true,
-  xslUrl,
-}: {
-  urls: (string | SitemapItemLoose)[];
-  targetFolder: string;
-  hostname?: string;
-  sitemapName?: string;
-  sitemapSize?: number;
-  gzip?: boolean;
-  xslUrl?: string;
-}): Promise<boolean> {
-  const indexStream = new SitemapIndexStream({ xslUrl });
-
-  try {
-    const stats = await statPromise(targetFolder);
-    if (!stats.isDirectory()) {
-      throw new UndefinedTargetFolder();
-    }
-  } catch (e) {
-    throw new UndefinedTargetFolder();
-  }
-
-  const indexWS = createWriteStream(
-    targetFolder + '/' + sitemapName + '-index.xml'
-  );
-  indexStream.pipe(indexWS);
-  const smPromises = chunk(urls, sitemapSize).map(
-    (chunk: (string | SitemapItemLoose)[], idx): Promise<boolean> => {
-      return new Promise((resolve, reject): void => {
-        const extension = '.xml' + (gzip ? '.gz' : '');
-        const filename = sitemapName + '-' + idx + extension;
-        indexStream.write(new URL(filename, hostname).toString());
-
-        const ws = createWriteStream(targetFolder + '/' + filename);
-        const sms = new SitemapStream({ hostname, xslUrl });
-        let pipe: Writable;
-        if (gzip) {
-          pipe = sms.pipe(createGzip()).pipe(ws);
-        } else {
-          pipe = sms.pipe(ws);
-        }
-        chunk.forEach((smi) => sms.write(smi));
-        sms.end();
-        pipe.on('finish', () => resolve(true));
-        pipe.on('error', (e) => reject(e));
-      });
-    }
-  );
-  return Promise.all(smPromises).then(() => {
-    indexStream.end();
-    return true;
-  });
-}
-
 type getSitemapStream = (
   i: number
 ) => [IndexItem | string, SitemapStream, WriteStream];
-/** @deprecated */
-type getSitemapStreamDeprecated = (
-  i: number
-) => [IndexItem | string, SitemapStream];
 
 export interface SitemapAndIndexStreamOptions
   extends SitemapIndexStreamOptions {
@@ -167,30 +77,15 @@ export interface SitemapAndIndexStreamOptions
   limit?: number;
   getSitemapStream: getSitemapStream;
 }
-export interface SitemapAndIndexStreamOptionsDeprecated
-  extends SitemapIndexStreamOptions {
-  level?: ErrorLevel;
-  limit?: number;
-  getSitemapStream: getSitemapStreamDeprecated;
-}
 // const defaultSIStreamOpts: SitemapAndIndexStreamOptions = {};
 export class SitemapAndIndexStream extends SitemapIndexStream {
   private i: number;
-  private getSitemapStream: getSitemapStream | getSitemapStreamDeprecated;
+  private getSitemapStream: getSitemapStream;
   private currentSitemap: SitemapStream;
   private currentSitemapPipeline?: WriteStream;
   private idxItem: IndexItem | string;
   private limit: number;
-  /**
-   * @deprecated this version does not properly wait for everything to write before resolving
-   * pass a 3rd param in your return from getSitemapStream that is the writeable stream
-   * to remove this warning
-   */
-  constructor(opts: SitemapAndIndexStreamOptionsDeprecated);
-  constructor(opts: SitemapAndIndexStreamOptions);
-  constructor(
-    opts: SitemapAndIndexStreamOptions | SitemapAndIndexStreamOptionsDeprecated
-  ) {
+  constructor(opts: SitemapAndIndexStreamOptions) {
     opts.objectMode = true;
     super(opts);
     this.i = 0;